service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
boto3==1.35.36
botocore==1.35.36
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.2.0
